pyproject.toml
[project]
name = "luhub-backend"
version = "0.1.0"
description = "API para a plataforma LuHub"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.15.1",
    "asyncpg>=0.30.0",
    "bcrypt>=4.3.0",
    "beanie>=1.29.0",
    "cryptography>=44.0.1",
    "fastapi>=0.115.10",
    "fastapi-cache>=0.1.0",
    "fastapi-cache2>=0.2.2",
    "fastapi-sqlalchemy>=0.2.1",
    "geopy[aiohttp]>=2.4.1",
    "httpx>=0.28.1",
    "ipython>=9.0.0",
    "itsdangerous>=2.2.0",
    "pandas>=2.2.3",
    "passlib>=1.7.4",
    "psycopg2-binary>=2.9.10",
    "pydantic>=2.10.6",
    "pydantic-settings>=2.8.1",
    "python-decouple>=3.8",
    "python-jose>=3.4.0",
    "python-multipart>=0.0.20",
    "requests>=2.32.3",
    "sqlalchemy>=2.0.38",
    "uvicorn>=0.34.0",
]
